Move symlink creation to post-upgrade script.
Apparently, and in hindsight unsurprisingly, the pre-upgrade script
runs *BEFORE* the files are switched over from the previous version of
the package to the new version of the package. So, any node upgrading
from 0.4.0 straight to 0.4.8 without going through 0.4.7 would have
pre-upgrade create the /data symlink, only for the file change logic
to then wipe it out.

By moving symlink creation to post-upgrade time, I expect the /data
symlink to survive the upgrade process.